backend/requirements.txt
flask 
requests
python-dotenv
langchain-google-genai
langchain-community
Flask-SQLAlchemy 
psycopg2-binary
redis
rq